mod7-streamlit/ejemplos/fifa_scouting_app/requirements.txt
fonttools==4.22.0
kiwisolver==1.0.1
matplotlib==3.6.2
matplotlib-inline 
pandas~=1.5.0
Pillow
scipy==1.11.4
streamlit~=1.25.0
highlight_text~=0.2
mplsoccer~=1.2.2